+++ app/src/main/java/com/tajmoti/tulip/ui/streams/StreamsAdapter.kt
@@ -1,13 +1,8 @@
 package com.tajmoti.tulip.ui.streams
 
-import android.text.Spannable
-import android.text.Spanned
-import android.text.style.StrikethroughSpan
-import android.widget.TextView
 import com.tajmoti.tulip.BaseAdapter
 import com.tajmoti.tulip.R
 import com.tajmoti.tulip.databinding.ItemStreamBinding
-
 
 class StreamsAdapter(
     val downloadCallback: (UnloadedVideoStreamRef) -> Unit
@@ -15,27 +10,9 @@ class StreamsAdapter(
 
     override fun onBindViewHolder(vh: Holder<ItemStreamBinding>, item: UnloadedVideoStreamRef) {
         val string = "#${vh.adapterPosition + 1}: ${item.info.serviceName}"
-        setTextConsiderBrokenLink(item.info.working, vh, string)
+        vh.binding.root.text = string
         vh.binding.root.setCompoundDrawablesRelativeWithIntrinsicBounds(getDrawable(item), 0, 0, 0)
         vh.binding.root.setOnLongClickListener { downloadCallback(item); true }
-    }
-
-    private fun setTextConsiderBrokenLink(
-        working: Boolean,
-        vh: Holder<ItemStreamBinding>,
-        string: String
-    ) {
-        if (working) {
-            vh.binding.root.text = string
-        } else {
-            vh.binding.root.setStrikethroughText(string)
-        }
-    }
-
-    private fun TextView.setStrikethroughText(string: String) {
-        setText(string, TextView.BufferType.SPANNABLE)
-        val spannable = text as Spannable
-        spannable.setSpan(STRIKE_THROUGH_SPAN, 0, string.length, Spanned.SPAN_EXCLUSIVE_EXCLUSIVE)
     }
 
     private fun getDrawable(ref: UnloadedVideoStreamRef): Int {
@@ -44,9 +21,5 @@ class StreamsAdapter(
         } else {
             R.drawable.ic_baseline_web_24
         }
-    }
-
-    companion object {
-        private val STRIKE_THROUGH_SPAN = StrikethroughSpan()
     }
 }