+++ app/src/main/java/com/tajmoti/tulip/ui/streams/StreamsViewModel.kt
@@ -1,15 +1,11 @@
 package com.tajmoti.tulip.ui.streams
 
 import android.net.Uri
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.Transformations
-import androidx.lifecycle.ViewModel
+import androidx.lifecycle.*
 import com.tajmoti.libtvprovider.Episode
 import com.tajmoti.libtvprovider.Streamable
 import com.tajmoti.libtvprovider.TvItem
 import com.tajmoti.libtvprovider.VideoStreamRef
-import com.tajmoti.libtvvideoextractor.VideoLinkExtractor
 import com.tajmoti.tulip.model.StreamableInfo
 import com.tajmoti.tulip.model.key.StreamableKey
 import com.tajmoti.tulip.service.StreamExtractorService
@@ -17,18 +13,19 @@ import com.tajmoti.tulip.service.TvDataService
 import com.tajmoti.tulip.service.VideoDownloadService
 import com.tajmoti.tulip.ui.performStatefulOneshotOperation
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.CancellationException
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
 class StreamsViewModel @Inject constructor(
     private val tvDataService: TvDataService,
-    private val linkExtractor: VideoLinkExtractor,
     private val downloadService: VideoDownloadService,
     private val extractionService: StreamExtractorService
 ) : ViewModel() {
     private val _state = MutableLiveData<State>(State.Idle)
-    private val _streamableName = MutableLiveData<String?>()
-    private val _directLoadingState = MutableLiveData<DirectStreamLoading?>()
+    private val _linkLoadingState = MutableLiveData<LinkLoadingState>(LinkLoadingState.Idle)
 
     /**
      * Loading state of the video stream links.
@@ -38,13 +35,21 @@ class StreamsViewModel @Inject constructor(
     /**
      * Name of the item, which the streams belong to.
      */
-    val streamableName: LiveData<String?> = _streamableName
+    val streamableName: LiveData<String?> = Transformations.map(streamLoadingState) {
+        val streamable = when (it) {
+            is State.Loading -> it.info.streamable
+            is State.Success -> it.info.streamable
+            is State.Error -> it.info?.streamable ?: return@map null
+            else -> return@map null
+        }
+        streamableToName(streamable)
+    }
 
     /**
      * Whether the stream links are being loaded right now.
      */
     val loading = Transformations.map(streamLoadingState)
-    { it is State.Loading }
+    { it is State.Idle || it is State.Preparing || it is State.Loading }
 
     /**
      * True when loading is finished, but no streams were found.
@@ -53,62 +58,92 @@ class StreamsViewModel @Inject constructor(
     { it is State.Success && it.streams.isEmpty() }
 
     /**
-     * If the user clicked a stream, which can be converted into a direct stream link,
-     * this is the state of that conversion. Null if no conversion was ever done.
+     * After the user clicks a link, this is the state of the link loading.
      */
-    val directStreamLoadingState: LiveData<DirectStreamLoading?> = _directLoadingState
+    val linkLoadingState: LiveData<LinkLoadingState?> = _linkLoadingState
 
     /**
-     * Whether a direct link conversion is currently being performed.
+     * Whether a stream loading or direct link conversion is currently being performed.
      */
-    val loadingDirectLink: LiveData<Boolean> = Transformations.map(directStreamLoadingState)
-    { it != null && it is DirectStreamLoading.Loading }
+    val loadingStreamOrDirectLink = Transformations.map(_linkLoadingState) {
+        it is LinkLoadingState.Loading || it is LinkLoadingState.LoadingDirect
+    }
+
+    /**
+     * Represents the link preparation and playing of the last clicked link.
+     */
+    private var linkExtractionJob: Job? = null
 
 
     /**
      * Search for a TV show or a movie.
      */
     fun fetchStreams(info: StreamableKey) {
-        performStatefulOneshotOperation(_state, State.Loading, State.Idle) {
+        performStatefulOneshotOperation(_state, State.Preparing, State.Idle) {
             fetchStreamsToState(info)
         }
     }
 
     /**
-     * Try to get direct link to video.
+     * The user has clicked a link, it needs to be resolved and played.
      */
-    fun fetchStreamDirect(ref: VideoStreamRef, download: Boolean) {
-        performStatefulOneshotOperation(
-            _directLoadingState,
-            DirectStreamLoading.Loading(ref, download),
-            null
-        ) { fetchStreamsToResult(ref, download) }
+    fun onStreamClicked(stream: UnloadedVideoStreamRef, download: Boolean) {
+        val info = stream.info
+        linkExtractionJob?.cancel()
+        linkExtractionJob = viewModelScope.launch {
+            try {
+                _linkLoadingState.value = when (info) {
+                    is VideoStreamRef.Resolved ->
+                        processResolvedLink(stream, info, download)
+                    is VideoStreamRef.Unresolved ->
+                        processUnresolvedLink(stream, info, download)
+                }
+            } catch (e: CancellationException) {
+                _linkLoadingState.value = LinkLoadingState.Idle
+            }
+        }
     }
-
-    private suspend fun fetchStreamsToResult(
-        ref: VideoStreamRef,
-        download: Boolean
-    ): DirectStreamLoading {
-        val result = linkExtractor.extractVideoLink(ref.url)
-            .getOrElse { return DirectStreamLoading.Failed(ref, it) }
-        return DirectStreamLoading.Success(ref, download, result)
-    }
-
 
     /**
-     * Download the video from the URL.
+     * Converts a redirect to a streaming page to the actual URL
+     * and passes it to [processResolvedLink].
      */
-    fun downloadVideo(link: String) {
+    private suspend fun processUnresolvedLink(
+        ref: UnloadedVideoStreamRef,
+        info: VideoStreamRef.Unresolved,
+        download: Boolean
+    ): LinkLoadingState {
+        _linkLoadingState.value = LinkLoadingState.Loading(info, download)
+        val resolvedUrl = extractionService.resolveStream(info)
+            .getOrElse { return LinkLoadingState.Error }
+        return processResolvedLink(ref, resolvedUrl, download)
+    }
+
+    private suspend fun processResolvedLink(
+        ref: UnloadedVideoStreamRef,
+        info: VideoStreamRef.Resolved,
+        download: Boolean
+    ): LinkLoadingState {
+        if (!ref.linkExtractionSupported)
+            return LinkLoadingState.DirectLinkUnsupported(info, download)
+        _linkLoadingState.value = LinkLoadingState.LoadingDirect(info, download)
+        val result = extractionService.extractVideoLink(info)
+            .getOrElse { return LinkLoadingState.Error }
+        if (download)
+            downloadVideo(result)
+        return LinkLoadingState.LoadedDirect(info, download, result)
+    }
+
+    private fun downloadVideo(link: String) {
         val state = streamLoadingState.value as State.Success
-        downloadService.downloadFileToFiles(Uri.parse(link), state.streamable)
+        downloadService.downloadFileToFiles(Uri.parse(link), state.info)
     }
 
     private suspend fun fetchStreamsToState(info: StreamableKey): State {
         val streamable = tvDataService.getStreamable(info)
-            .getOrElse { return State.Error }
-        _streamableName.value = streamableToName(streamable.streamable)
+            .getOrElse { return State.Error(null) }
         val result = extractionService.fetchStreams(streamable)
-            .getOrElse { return State.Error }
+            .getOrElse { return State.Error(streamable) }
         return State.Success(result.info, result.streams)
     }
 
@@ -120,36 +155,78 @@ class StreamsViewModel @Inject constructor(
     }
 
     sealed class State {
+        /**
+         * Marker state before the loading is started.
+         */
         object Idle : State()
-        object Loading : State()
+
+        /**
+         * Loading item information from the DB.
+         */
+        object Preparing : State()
+
+        /**
+         * Loading streams from the TV provider website.
+         */
+        data class Loading(val info: StreamableInfo) : State()
+
+        /**
+         * Streamable item loaded successfully.
+         */
         data class Success(
-            val streamable: StreamableInfo,
+            val info: StreamableInfo,
             val streams: List<UnloadedVideoStreamRef>
         ) : State()
 
-        object Error : State()
+        /**
+         * Error during loading of the item.
+         * Null if loading from the DB failed.
+         */
+        data class Error(val info: StreamableInfo?) : State()
 
         val success: Boolean
             get() = this is Success
     }
 
-    sealed class DirectStreamLoading {
-        abstract val video: VideoStreamRef
+    sealed class LinkLoadingState {
+        /**
+         * No link clicked yet.
+         */
+        object Idle : LinkLoadingState()
 
+        /**
+         * The streaming page URL is being resolved.
+         */
         data class Loading(
-            override val video: VideoStreamRef,
+            val stream: VideoStreamRef.Unresolved,
             val download: Boolean
-        ) : DirectStreamLoading()
+        ) : LinkLoadingState()
 
-        data class Success(
-            override val video: VideoStreamRef,
+        /**
+         * Direct link extraction is not supported for the clicked streaming site.
+         */
+        data class DirectLinkUnsupported(
+            val stream: VideoStreamRef.Resolved,
+            val download: Boolean
+        ) : LinkLoadingState()
+
+        /**
+         * A direct video link is being extracted.
+         */
+        data class LoadingDirect(
+            val stream: VideoStreamRef.Resolved,
+            val download: Boolean
+        ) : LinkLoadingState()
+
+        /**
+         * A direct video link was extracted successfully.
+         */
+        data class LoadedDirect(
+            val stream: VideoStreamRef.Resolved,
             val download: Boolean,
             val directLink: String
-        ) : DirectStreamLoading()
+        ) : LinkLoadingState()
 
-        data class Failed(
-            override val video: VideoStreamRef,
-            val error: Throwable
-        ) : DirectStreamLoading()
+        object Error : LinkLoadingState()
     }
 }