+++ libtvprovider/src/main/java/com/tajmoti/libtvprovider/VideoStreamRef.kt
@@ -1,7 +1,35 @@
 package com.tajmoti.libtvprovider
 
-data class VideoStreamRef(
-    val serviceName: String,
-    val url: String,
-    val working: Boolean
-)
+sealed class VideoStreamRef {
+    /**
+     * Name of the hosting site (informative).
+     */
+    abstract val serviceName: String
+
+    /**
+     * URL to the streaming page (or a redirect in case this is [Unresolved].
+     */
+    abstract val url: String
+
+    /**
+     * A redirect to a streaming page.
+     * Needs a HTML GET request to resolve the redirect.
+     */
+    data class Unresolved(
+        override val serviceName: String,
+        override val url: String,
+    ) : VideoStreamRef() {
+
+        fun asResolved(resolvedUrl: String): Resolved {
+            return Resolved(serviceName, resolvedUrl)
+        }
+    }
+
+    /**
+     * A real URL to the video streaming page.
+     */
+    data class Resolved(
+        override val serviceName: String,
+        override val url: String
+    ) : VideoStreamRef()
+}